move service test files
