+++ src/app/services/actions/permissions-panel-action-creator.spec.ts
@@ -1,5 +1,5 @@
-import { PERMISSIONS_PANEL_OPEN } from '../../../app/services/redux-constants';
-import { togglePermissionsPanel } from '../../../app/services/actions/permissions-panel-action-creator';
+import { PERMISSIONS_PANEL_OPEN } from '../redux-constants';
+import { togglePermissionsPanel } from './permissions-panel-action-creator';
 
 describe('Tests permissions panel action creator', () => {
   it('tests permissions panel action creator', () => {
