+++ src/app/services/reducers/adaptive-cards-reducer.spec.ts
