+++ src/app/services/reducers/terms-of-use-reducer.spec.ts
